migrated all utility functions
