+++ src/constants.py
@@ -1,0 +1,17 @@
+import os
+
+# for gitpod environment
+from dotenv import load_dotenv
+load_dotenv()
+
+TOKEN = os.environ.get("IGCSEBOT_TOKEN")
+LINK = "" # os.environ.get("MONGO_LINK")
+
+GUILD_ID = 1111128710133854289 # 576460042774118420
+LOG_CHANNEL_ID = 1152467554321842257 # 1017792876584906782
+CREATE_DM_CHANNEL_ID = 895961641219407923
+
+SHOULD_LOG_ALL = False
+
+SESSION_ROLES = {782273117321166888, 782423045578948618, 782423808271056917, 853135207862370324, 853135509488009216, 853135836597583902, 931662314074169364, 921725505051459664, 921725841677901905, 698291687981580308}
+SUBJECT_ROLES = {688354722251276308, 688355303808303170, 871702273640787988, 685837416443281493, 685837450895032336, 685837475939221770, 667769546475700235, 688357525984509984, 685837363003523097, 685836740774330378, 677411157434171404, 688357284920819717, 668927421512155142, 688359807203278986, 685837255138607182, 688356986798211074, 688362197096988688, 688356500506148884, 685836973004423169, 685837280396705803, 688360519735967766, 688362250951852085, 685837171491471401, 685836834005319721, 868056692942835742, 932660913054580766, 688361654852780048, 871587296330260570, 685836866166980661, 688361164161548361, 688361596149563431, 685836889990627335, 871587728737849344, 886884445657890826, 853588511252545566, 853587095317643264, 853588416774930432, 853586114380300298, 853586292435451904, 853586450141413416, 883190243623333888, 883190176644468758, 881366711645921310, 868323020769493034, 883190289211228170, 868324666102661140, 788344643367469057}
