+++ src/on_command_error.py
@@ -1,0 +1,12 @@
+from constants import LINK, GUILD_ID, LOG_CHANNEL_ID
+from bot import bot
+
+@bot.event
+async def on_command_error(ctx, exception):
+    if isinstance(exception, commands.CommandNotFound):
+        return
+    description = f"Channel: {ctx.channel.mention}\nUser: {ctx.author.mention}\nGuild: {ctx.guild.name} ({ctx.guild.id})\n\nError:\n```{''.join(traceback.format_exception(exception, exception, exception.__traceback__))}```"
+    embed = discord.Embed(title="An Exception Occured", description=description)
+    igcse = await bot.fetch_guild(GUILD_ID)
+    logs = await igcse.fetch_channel(LOG_CHANNEL_ID)
+    await logs.send(embed=embed)
