+++ src/roles.py
@@ -1,0 +1,25 @@
+from bot import discord
+
+async def has_role(member: discord.Member, role_name: str):
+    roles = [role.name.lower() for role in member.roles]
+    for role in roles:
+        if role_name.lower() in role:
+            return True
+    return False
+
+# unused function
+async def get_role(role_name: str):
+    guild = bot.get_guild(GUILD_ID)
+    role = discord.utils.get(guild.roles, name = role_name)
+    return role
+
+async def is_moderator(member: discord.Member):
+    roles = [role.id for role in member.roles]
+    if 578170681670369290 in roles or 784673059906125864 in roles:  # r/igcse moderator role ids
+        return True
+    elif member.guild_permissions.administrator:
+        return True
+    return False
+
+async def is_server_booster(member: discord.Member):
+    return await has_role(member, "Server Booster")
